requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
ipykernel==6.23.2
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.1
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
oauthlib==3.2.2
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.5.3
pluggy==1.0.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pytest==7.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
stack-data==0.6.2
tomli==2.0.1
tornado==6.3.2
traitlets==5.9.0
tweepy==4.14.0
typing_extensions==4.6.3
urllib3==2.0.3
wcwidth==0.2.6
Werkzeug==2.3.6
zipp==3.15.0
